pyproject.toml
[project]
name = "comfyui-sync-lipsync-node"
description = ""
version = "1.0.1"
license = {file = "LICENSE"}
dependencies = ["requests==2.32.3", "syncsdk==0.1.9", "protobuf<6.0.0,>=3.20.3"]

[project.urls]
Repository = "https://github.com/wasilone11/comfyui-sync-lipsync-node"
#  Used by Comfy Registry https://registry.comfy.org

[tool.comfy]
PublisherId = "wasilone-sync"
DisplayName = "comfyui-sync-lipsync-node"
Icon = ""
includes = []
